+++ iterableapi/src/main/java/com/iterable/iterableapi/IterableActivityMonitor.java
@@ -119,6 +119,12 @@ public class IterableActivityMonitor {
     }
 
     public void addCallback(@NonNull AppStateCallback callback) {
+        // Don't insert again if the same callback already exists
+        for (WeakReference<AppStateCallback> existingCallback : callbacks) {
+            if (existingCallback.get() == callback) {
+                return;
+            }
+        }
         callbacks.add(new WeakReference<>(callback));
     }